worker to be added

+++ app/src/main/java/com/hitmeows/weathercat/features/weather/data/remote/dto/air_pollution/Components.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Components(
     val co: Double,
-    val no: Int,
+    val no: Double,
     val no2: Double,
     val o3: Double,
     val so2: Double,
