+++ app/src/main/java/com/hitmeows/weathercat/features/weather/data/remote/dto/air_pollution/Coord.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Coord(
-    val lon: Int,
-    val lat: Int
+    val lon: Double,
+    val lat: Double
 )
