+++ app/src/main/java/com/hitmeows/weathercat/features/weather/data/remote/dto/all_weather/AllWeatherDto.kt
@@ -1,8 +1,12 @@
 package com.hitmeows.weathercat.features.weather.data.remote.dto.all_weather
 
 
+import com.hitmeows.weathercat.features.weather.data.local.Coordinates
+import com.hitmeows.weathercat.features.weather.data.local.entities.DailyWeather
+import com.hitmeows.weathercat.features.weather.data.local.entities.HourlyWeather
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlin.math.roundToInt
 
 @Serializable
 data class AllWeatherDto(
@@ -10,7 +14,7 @@ data class AllWeatherDto(
     val lon: Double,
     val timezone: String,
     @SerialName("timezone_offset")
-    val timezoneOffset: Int,
+    val timezoneOffset: Long,
     val hourly: List<Hourly>,
     val daily: List<Daily>
 )
